+++ TypeScript.ContractGenerator.Tests/Files/Options/pluralize-default.ts
@@ -5,12 +5,12 @@ export type EnumContainingRootType = {
     explicitEnum: ExplicitEnum;
 };
 export type DefaultEnum = 'A' | 'B';
-export const DefaultEnumItems = {
+export const DefaultEnums = {
     ['A']: ('A') as DefaultEnum,
     ['B']: ('B') as DefaultEnum,
 };
 export type ExplicitEnum = 'C' | 'D';
-export const ExplicitEnumItems = {
+export const ExplicitEnums = {
     ['C']: ('C') as ExplicitEnum,
     ['D']: ('D') as ExplicitEnum,
 };