+++ TypeScript.ContractGenerator.Tests/OptionsTests.cs
@@ -28,7 +28,7 @@ namespace SkbKontur.TypeScript.ContractGenerator.Tests
         [TestCase(false, "optional-properties-disabled")]
         public void OptionalPropertiesTest(bool optionalPropertiesEnabled, string expectedFileName)
         {
-            var generatedCode = GenerateCode(new TypeScriptGenerationOptions {EnableOptionalProperties = optionalPropertiesEnabled}, CustomTypeGenerator.Null, typeof(SingleNullablePropertyType)).Single();
+            var generatedCode = GenerateCode(new TypeScriptGenerationOptions {EnableOptionalProperties = optionalPropertiesEnabled, EnumGenerationMode = EnumGenerationMode.FixedStringsAndDictionary}, CustomTypeGenerator.Null, typeof(SingleNullablePropertyType)).Single();
             var expectedCode = GetExpectedCode($"Options/{expectedFileName}");
             generatedCode.Diff(expectedCode).ShouldBeEmpty();
         }
@@ -37,7 +37,7 @@ namespace SkbKontur.TypeScript.ContractGenerator.Tests
         [TestCase(false, "explicit-nullability-disabled")]
         public void ExplicitNullabilityTest(bool explicitNullabilityEnabled, string expectedFileName)
         {
-            var generatedCode = GenerateCode(new TypeScriptGenerationOptions {EnableExplicitNullability = explicitNullabilityEnabled}, CustomTypeGenerator.Null, typeof(ExplicitNullabilityRootType)).Single();
+            var generatedCode = GenerateCode(new TypeScriptGenerationOptions {EnableExplicitNullability = explicitNullabilityEnabled, EnumGenerationMode = EnumGenerationMode.FixedStringsAndDictionary}, CustomTypeGenerator.Null, typeof(ExplicitNullabilityRootType)).Single();
             var expectedCode = GetExpectedCode($"Options/{expectedFileName}");
             generatedCode.Diff(expectedCode).ShouldBeEmpty();
         }
@@ -46,7 +46,7 @@ namespace SkbKontur.TypeScript.ContractGenerator.Tests
         [TestCase(false, "global-nullable-disabled")]
         public void GlobalNullableTest(bool useGlobalNullable, string expectedFileName)
         {
-            var generatedCode = GenerateCode(new TypeScriptGenerationOptions {UseGlobalNullable = useGlobalNullable}, CustomTypeGenerator.Null, typeof(GlobalNullableRootType)).Single();
+            var generatedCode = GenerateCode(new TypeScriptGenerationOptions {UseGlobalNullable = useGlobalNullable, EnumGenerationMode = EnumGenerationMode.FixedStringsAndDictionary}, CustomTypeGenerator.Null, typeof(GlobalNullableRootType)).Single();
             var expectedCode = GetExpectedCode($"Options/{expectedFileName}");
             generatedCode.Diff(expectedCode).ShouldBeEmpty();
         }
@@ -55,7 +55,7 @@ namespace SkbKontur.TypeScript.ContractGenerator.Tests
         [TestCase("pluralize-custom", "Items")]
         public void PluralizeTest(string expectedFileName, string pluralizeSuffix)
         {
-            var options = TypeScriptGenerationOptions.Default;
+            var options = TestOptions;
             if (!string.IsNullOrEmpty(pluralizeSuffix))
                 options.Pluralize = s => s + pluralizeSuffix;
 
@@ -68,7 +68,7 @@ namespace SkbKontur.TypeScript.ContractGenerator.Tests
         [TestCase("nullability-optimistic", NullabilityMode.Optimistic)]
         public void NullabilityModeTest(string expectedFileName, NullabilityMode mode)
         {
-            var options = TypeScriptGenerationOptions.Default;
+            var options = TestOptions;
             options.NullabilityMode = mode;
 
             var generatedCode = GenerateCode(options, CustomTypeGenerator.Null, typeof(NullabilityModeRootType)).Single();
@@ -79,7 +79,7 @@ namespace SkbKontur.TypeScript.ContractGenerator.Tests
         [Test]
         public void TestNullableReferences()
         {
-            var options = new TypeScriptGenerationOptions {NullabilityMode = NullabilityMode.NullableReference};
+            var options = new TypeScriptGenerationOptions {NullabilityMode = NullabilityMode.NullableReference, EnumGenerationMode = EnumGenerationMode.FixedStringsAndDictionary};
             var generatedCode = GenerateCode(options, CustomTypeGenerator.Null, typeof(NullableReferenceType)).Single();
             var expectedCode = GetExpectedCode("Options/nullable-reference");
             generatedCode.Diff(expectedCode).ShouldBeEmpty();