+++ TypeScript.ContractGenerator/CodeDom/JavaScriptTypeChecker.cs
@@ -1,8 +1,12 @@
+using System;
+
 namespace SkbKontur.TypeScript.ContractGenerator.CodeDom
 {
     public enum JavaScriptTypeChecker
     {
+        [Obsolete("Will be removed in 2.0")]
         Flow,
+
         TypeScript
     }
 }